LPS-179960 Rename package
