+++ modules/apps/object/object-api/bnd.bnd
@@ -10,7 +10,6 @@ Export-Package:\
 	com.liferay.object.definition.notification.term.util,\
 	com.liferay.object.deployer,\
 	com.liferay.object.display.context,\
-	com.liferay.object.dsl,\
 	com.liferay.object.entry.util,\
 	com.liferay.object.exception,\
 	com.liferay.object.field.builder,\
@@ -23,6 +22,7 @@ Export-Package:\
 	com.liferay.object.field.util,\
 	com.liferay.object.filter.util,\
 	com.liferay.object.model,\
+	com.liferay.object.petra.sql.dsl,\
 	com.liferay.object.related.models,\
 	com.liferay.object.relationship.util,\
 	com.liferay.object.rest.context.path,\
